리뷰 평균 조회

+++ src/main/java/com/example/WebOrder/dto/MenuStatisticsDto.java
@@ -12,7 +12,7 @@ public class MenuStatisticsDto {
 
     //가장 별점이 좋은 메뉴
     private String bestRateMenu;
-    private Integer bestRateNum;
+    private Double bestRateNum;
 
 
     //가장 많이 주문된 메뉴
